refactoring runmat-hir for maintainability
