+++ crates/runmat-hir/src/inference/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod expr;
+pub(crate) mod function_outputs;
+pub(crate) mod function_vars;
+pub(crate) mod globals;
+pub(crate) mod shared;
