+++ crates/runmat-hir/src/lowering/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod ctx;
+pub(crate) mod expr;
+pub(crate) mod stmt;
+
+pub use ctx::lower;
